Address review comments: remove user_id_tensor and fix rope_max_seq_len
Remove unused user_id_tensor parameter from the entire call chain
(attention, decoder, model, generator). Revert rope_max_seq_len
inflation that corrupted RoPE position embeddings. Remove stale
kwargs.get() calls and revert int64 cast in 70B Galaxy generator.

Made-with: Cursor

+++ models/demos/llama3_70b_galaxy/tt/generator.py
@@ -395,9 +395,7 @@ class Generator(WarmupForwardMixin):
             sampled_tokens = ttnn.to_torch(ttnn.get_device_tensors(tt_sampled)[0]).to(torch.int32)
 
             # sampled_tokens has 32 entries ordered by slot.
-            # Cast to int64: ttnn.to_torch returns uint32 for token ids which doesn't
-            # support CPU fancy-indexing on PyTorch.
-            sampled_tensor = sampled_tokens[0, 0, 0, :].to(torch.int64)  # Shape: [32]
+            sampled_tensor = sampled_tokens[0, 0, 0, :]  # Shape: [32]
             output_toks = sampled_tensor[empty_slots]
 
             if tt_log_probs is not None:
